Upd: removed imwrite+ own file
small tweaks
