+++ module/retire/assets.py
@@ -6,7 +6,7 @@ from module.base.template import Template
 
 COMMON_SHIP_FILTER_DISABLE = Button(area={'cn': (714, 13, 757, 41), 'en': (706, 15, 758, 41), 'jp': (711, 13, 759, 42), 'tw': (712, 12, 759, 42)}, color={'cn': (127, 143, 183), 'en': (86, 99, 138), 'jp': (119, 131, 170), 'tw': (117, 132, 173)}, button={'cn': (714, 13, 757, 41), 'en': (706, 15, 758, 41), 'jp': (711, 13, 759, 42), 'tw': (712, 12, 759, 42)}, file={'cn': './assets/cn/retire/COMMON_SHIP_FILTER_DISABLE.png', 'en': './assets/en/retire/COMMON_SHIP_FILTER_DISABLE.png', 'jp': './assets/jp/retire/COMMON_SHIP_FILTER_DISABLE.png', 'tw': './assets/tw/retire/COMMON_SHIP_FILTER_DISABLE.png'})
 COMMON_SHIP_FILTER_ENABLE = Button(area={'cn': (713, 12, 758, 41), 'en': (707, 16, 758, 40), 'jp': (712, 13, 758, 42), 'tw': (712, 11, 759, 42)}, color={'cn': (214, 174, 130), 'en': (193, 149, 106), 'jp': (212, 171, 129), 'tw': (211, 170, 127)}, button={'cn': (713, 12, 758, 41), 'en': (707, 16, 758, 40), 'jp': (712, 13, 758, 42), 'tw': (712, 11, 759, 42)}, file={'cn': './assets/cn/retire/COMMON_SHIP_FILTER_ENABLE.png', 'en': './assets/en/retire/COMMON_SHIP_FILTER_ENABLE.png', 'jp': './assets/jp/retire/COMMON_SHIP_FILTER_ENABLE.png', 'tw': './assets/tw/retire/COMMON_SHIP_FILTER_ENABLE.png'})
-DOCK_AMOUNT = Button(area={'cn': (473, 17, 611, 41), 'en': (513, 15, 588, 42), 'jp': (473, 14, 584, 42), 'tw': (473, 17, 611, 41)}, color={'cn': (71, 74, 78), 'en': (44, 49, 54), 'jp': (46, 53, 61), 'tw': (71, 74, 78)}, button={'cn': (473, 17, 611, 41), 'en': (513, 15, 588, 42), 'jp': (473, 14, 584, 42), 'tw': (473, 17, 611, 41)}, file={'cn': './assets/cn/retire/DOCK_AMOUNT.png', 'en': './assets/en/retire/DOCK_AMOUNT.png', 'jp': './assets/jp/retire/DOCK_AMOUNT.png', 'tw': './assets/tw/retire/DOCK_AMOUNT.png'})
+DOCK_AMOUNT = Button(area={'cn': (473, 17, 611, 41), 'en': (476, 9, 611, 44), 'jp': (473, 14, 584, 42), 'tw': (473, 17, 611, 41)}, color={'cn': (71, 74, 78), 'en': (50, 54, 58), 'jp': (46, 53, 61), 'tw': (71, 74, 78)}, button={'cn': (473, 17, 611, 41), 'en': (476, 9, 611, 44), 'jp': (473, 14, 584, 42), 'tw': (473, 17, 611, 41)}, file={'cn': './assets/cn/retire/DOCK_AMOUNT.png', 'en': './assets/en/retire/DOCK_AMOUNT.png', 'jp': './assets/jp/retire/DOCK_AMOUNT.png', 'tw': './assets/tw/retire/DOCK_AMOUNT.png'})
 DOCK_CHECK = Button(area={'cn': (121, 14, 175, 39), 'en': (121, 17, 189, 39), 'jp': (232, 12, 308, 41), 'tw': (121, 13, 176, 39)}, color={'cn': (156, 171, 207), 'en': (112, 125, 162), 'jp': (88, 96, 128), 'tw': (149, 164, 200)}, button={'cn': (121, 14, 175, 39), 'en': (121, 17, 189, 39), 'jp': (232, 12, 308, 41), 'tw': (121, 13, 176, 39)}, file={'cn': './assets/cn/retire/DOCK_CHECK.png', 'en': './assets/en/retire/DOCK_CHECK.png', 'jp': './assets/jp/retire/DOCK_CHECK.png', 'tw': './assets/tw/retire/DOCK_CHECK.png'})
 DOCK_EMPTY = Button(area={'cn': (95, 347, 388, 378), 'en': (95, 318, 264, 339), 'jp': (96, 347, 252, 376), 'tw': (94, 347, 390, 379)}, color={'cn': (160, 154, 159), 'en': (106, 99, 106), 'jp': (159, 152, 156), 'tw': (163, 157, 162)}, button={'cn': (95, 347, 388, 378), 'en': (95, 318, 264, 339), 'jp': (96, 347, 252, 376), 'tw': (94, 347, 390, 379)}, file={'cn': './assets/cn/retire/DOCK_EMPTY.png', 'en': './assets/en/retire/DOCK_EMPTY.png', 'jp': './assets/jp/retire/DOCK_EMPTY.png', 'tw': './assets/tw/retire/DOCK_EMPTY.png'})
 DOCK_FILTER = Button(area={'cn': (1099, 5, 1193, 48), 'en': (1098, 4, 1194, 49), 'jp': (1101, 6, 1192, 46), 'tw': (1099, 6, 1193, 47)}, color={'cn': (70, 87, 127), 'en': (73, 90, 128), 'jp': (67, 84, 125), 'tw': (78, 96, 137)}, button={'cn': (1099, 5, 1193, 48), 'en': (1098, 4, 1194, 49), 'jp': (1101, 6, 1192, 46), 'tw': (1099, 6, 1193, 47)}, file={'cn': './assets/cn/retire/DOCK_FILTER.png', 'en': './assets/en/retire/DOCK_FILTER.png', 'jp': './assets/jp/retire/DOCK_FILTER.png', 'tw': './assets/tw/retire/DOCK_FILTER.png'})
@@ -14,7 +14,7 @@ DOCK_FILTER_CONFIRM = Button(area={'cn': (714, 613, 886, 671), 'en': (718, 618, 
 DOCK_FIRST_NPC = Button(area={'cn': (96, 111, 123, 123), 'en': (96, 111, 123, 123), 'jp': (96, 111, 123, 123), 'tw': (96, 111, 123, 123)}, color={'cn': (184, 150, 150), 'en': (184, 150, 150), 'jp': (184, 150, 150), 'tw': (184, 150, 150)}, button={'cn': (96, 111, 123, 123), 'en': (96, 111, 123, 123), 'jp': (96, 111, 123, 123), 'tw': (96, 111, 123, 123)}, file={'cn': './assets/cn/retire/DOCK_FIRST_NPC.png', 'en': './assets/cn/retire/DOCK_FIRST_NPC.png', 'jp': './assets/cn/retire/DOCK_FIRST_NPC.png', 'tw': './assets/cn/retire/DOCK_FIRST_NPC.png'})
 DOCK_SCROLL = Button(area={'cn': (1239, 76, 1248, 641), 'en': (1239, 76, 1248, 641), 'jp': (1237, 78, 1250, 628), 'tw': (1239, 76, 1248, 641)}, color={'cn': (47, 46, 37), 'en': (47, 46, 37), 'jp': (180, 156, 66), 'tw': (47, 46, 37)}, button={'cn': (1239, 76, 1248, 641), 'en': (1239, 76, 1248, 641), 'jp': (1237, 78, 1250, 628), 'tw': (1239, 76, 1248, 641)}, file={'cn': './assets/cn/retire/DOCK_SCROLL.png', 'en': './assets/en/retire/DOCK_SCROLL.png', 'jp': './assets/jp/retire/DOCK_SCROLL.png', 'tw': './assets/tw/retire/DOCK_SCROLL.png'})
 DOCK_SELECTED = Button(area={'cn': (582, 662, 647, 685), 'en': (597, 658, 655, 691), 'jp': (603, 662, 655, 685), 'tw': (582, 662, 647, 685)}, color={'cn': (75, 75, 83), 'en': (64, 62, 73), 'jp': (84, 83, 92), 'tw': (75, 75, 83)}, button={'cn': (582, 662, 647, 685), 'en': (597, 658, 655, 691), 'jp': (603, 662, 655, 685), 'tw': (582, 662, 647, 685)}, file={'cn': './assets/cn/retire/DOCK_SELECTED.png', 'en': './assets/en/retire/DOCK_SELECTED.png', 'jp': './assets/jp/retire/DOCK_SELECTED.png', 'tw': './assets/tw/retire/DOCK_SELECTED.png'})
-EMPTY_ENHANCE_SLOT_PLUS = Button(area={'cn': (737, 402, 773, 437), 'en': (747, 402, 782, 437), 'jp': (737, 402, 773, 437), 'tw': (737, 402, 773, 437)}, color={'cn': (46, 46, 46), 'en': (38, 38, 38), 'jp': (46, 46, 46), 'tw': (46, 46, 46)}, button={'cn': (737, 402, 773, 437), 'en': (747, 402, 782, 437), 'jp': (737, 402, 773, 437), 'tw': (737, 402, 773, 437)}, file={'cn': './assets/cn/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'en': './assets/en/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'jp': './assets/jp/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'tw': './assets/tw/retire/EMPTY_ENHANCE_SLOT_PLUS.png'})
+EMPTY_ENHANCE_SLOT_PLUS = Button(area={'cn': (737, 402, 773, 437), 'en': (737, 402, 773, 437), 'jp': (737, 402, 773, 437), 'tw': (737, 402, 773, 437)}, color={'cn': (46, 46, 46), 'en': (46, 46, 46), 'jp': (46, 46, 46), 'tw': (46, 46, 46)}, button={'cn': (737, 402, 773, 437), 'en': (737, 402, 773, 437), 'jp': (737, 402, 773, 437), 'tw': (737, 402, 773, 437)}, file={'cn': './assets/cn/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'en': './assets/en/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'jp': './assets/jp/retire/EMPTY_ENHANCE_SLOT_PLUS.png', 'tw': './assets/tw/retire/EMPTY_ENHANCE_SLOT_PLUS.png'})
 ENHANCE_CONFIRM = Button(area={'cn': (1126, 602, 1256, 645), 'en': (1226, 605, 1262, 641), 'jp': (1126, 601, 1257, 646), 'tw': (1215, 606, 1253, 641)}, color={'cn': (203, 149, 81), 'en': (212, 167, 100), 'jp': (189, 139, 78), 'tw': (216, 169, 99)}, button={'cn': (1126, 602, 1256, 645), 'en': (1226, 605, 1262, 641), 'jp': (1126, 601, 1257, 646), 'tw': (1215, 606, 1253, 641)}, file={'cn': './assets/cn/retire/ENHANCE_CONFIRM.png', 'en': './assets/en/retire/ENHANCE_CONFIRM.png', 'jp': './assets/jp/retire/ENHANCE_CONFIRM.png', 'tw': './assets/tw/retire/ENHANCE_CONFIRM.png'})
 ENHANCE_FILLED = Button(area={'cn': (728, 440, 781, 454), 'en': (728, 440, 781, 454), 'jp': (722, 387, 789, 400), 'tw': (728, 440, 781, 454)}, color={'cn': (156, 138, 127), 'en': (156, 138, 127), 'jp': (146, 153, 211), 'tw': (156, 138, 127)}, button={'cn': (728, 440, 781, 454), 'en': (728, 440, 781, 454), 'jp': (722, 387, 789, 400), 'tw': (728, 440, 781, 454)}, file={'cn': './assets/cn/retire/ENHANCE_FILLED.png', 'en': './assets/en/retire/ENHANCE_FILLED.png', 'jp': './assets/jp/retire/ENHANCE_FILLED.png', 'tw': './assets/tw/retire/ENHANCE_FILLED.png'})
 ENHANCE_RECOMMEND = Button(area={'cn': (959, 602, 1089, 645), 'en': (1057, 606, 1091, 640), 'jp': (965, 605, 1085, 641), 'tw': (1050, 607, 1082, 639)}, color={'cn': (87, 134, 194), 'en': (106, 151, 215), 'jp': (86, 136, 201), 'tw': (106, 151, 212)}, button={'cn': (959, 602, 1089, 645), 'en': (1057, 606, 1091, 640), 'jp': (965, 605, 1085, 641), 'tw': (1050, 607, 1082, 639)}, file={'cn': './assets/cn/retire/ENHANCE_RECOMMEND.png', 'en': './assets/en/retire/ENHANCE_RECOMMEND.png', 'jp': './assets/jp/retire/ENHANCE_RECOMMEND.png', 'tw': './assets/tw/retire/ENHANCE_RECOMMEND.png'})
